fix: address PR #7 review feedback
- Move Video Player from the Transform resource to a new singular `Widget`
  resource: handler relocated to operations/widget/, controls split into
  widget.fields.ts, shared Public ID/Type fields broadened to both resources,
  tests and architecture docs updated, handler re-keyed `widget:videoPlayer`.
- Revert the codex `node` identifier to `n8n-nodes-base.cloudinary` to match
  the n8n codex-file spec and published community-node practice.

Also bundles accompanying working-tree changes (utils refactor, removal of the
updateDisplayName operation, multiStep→combineTransformations handler key, and
package metadata cleanup).
