+++ nodes/Cloudinary/Cloudinary.node.json
@@ -1,5 +1,5 @@
 {
-	"node": "n8n-nodes-cloudinary.cloudinary",
+	"node": "n8n-nodes-base.cloudinary",
 	"nodeVersion": "1.0",
 	"codexVersion": "1.0",
 	"categories": ["Data & Storage", "Marketing & Content"],
