+++ nodes/Cloudinary/descriptions/transform.fields.ts
@@ -7,6 +7,9 @@ import { INodeProperties } from 'n8n-workflow';
 // fields). Transform-specific knobs that have no API counterpart use `transform`-/
 // op-prefixed names to stay unambiguous.
 
+// Every op that references a stored asset by public_id — the Transform ops plus the
+// Widgets `videoPlayer` op (Widgets resource), which shares the Public ID + Type
+// identity fields below even though its own controls live in widget.fields.ts.
 const ALL_TRANSFORM_OPS = [
 	'optimizeImage',
 	'resizeImage',
@@ -16,7 +19,7 @@ const ALL_TRANSFORM_OPS = [
 	'trimVideo',
 	'videoThumbnail',
 	'customTransformation',
-	'multiStep',
+	'combineTransformations',
 	'videoPlayer',
 ];
 
@@ -45,7 +48,7 @@ export const transformFields: INodeProperties[] = [
 			'Public ID of the asset, including any folder path and without the file extension',
 		displayOptions: {
 			show: {
-				resource: ['transform'],
+				resource: ['transform', 'widget'],
 				operation: ALL_TRANSFORM_OPS,
 			},
 		},
@@ -74,7 +77,7 @@ export const transformFields: INodeProperties[] = [
 		description: 'Cloudinary delivery type — in almost all cases use "Upload" (public assets). Becomes the URL path segment (e.g. image/upload/sample, video/fetch/&lt;URL&gt;). IMPORTANT: "Private" and "Authenticated" assets — and, depending on account settings, social profile sources (Facebook, Twitter, Gravatar) — require a signed URL, which this node does not generate yet, so those delivery URLs will fail to load. Leave as "Upload" unless you specifically need another type.',
 		displayOptions: {
 			show: {
-				resource: ['transform'],
+				resource: ['transform', 'widget'],
 				operation: ALL_TRANSFORM_OPS,
 			},
 		},
@@ -126,7 +129,7 @@ export const transformFields: INodeProperties[] = [
 			{ name: 'Scale (Exact)', value: 'scale', description: 'Force exact dimensions; may distort if both are set' },
 		],
 		default: 'limit',
-		description: 'How the image is fitted to the requested dimensions. Note: Resize does not auto-optimize — chain Optimize Image after it, or use Multi-Step, to add f_auto/q_auto.',
+		description: 'How the image is fitted to the requested dimensions. Note: Resize does not auto-optimize — chain Optimize Image after it, or use Combine Transformations, to add f_auto/q_auto.',
 		displayOptions: {
 			show: { resource: ['transform'], operation: ['resizeImage'] },
 		},
@@ -200,7 +203,7 @@ export const transformFields: INodeProperties[] = [
 			{ name: 'Center', value: 'center', description: 'Center the crop on the middle of the image' },
 		],
 		default: 'auto',
-		description: 'Which part of the image to keep in focus when cropping. Note: Crop does not auto-optimize — chain Optimize Image after it, or use Multi-Step, to add f_auto/q_auto.',
+		description: 'Which part of the image to keep in focus when cropping. Note: Crop does not auto-optimize — chain Optimize Image after it, or use Combine Transformations, to add f_auto/q_auto.',
 		displayOptions: {
 			show: { resource: ['transform'], operation: ['cropImage'] },
 		},
@@ -401,255 +404,7 @@ export const transformFields: INodeProperties[] = [
 		},
 	},
 
-	// ── Video Player ─────────────────────────────────────────────────────────
-	{
-		displayName: 'Autoplay Mode',
-		name: 'playerAutoplayMode',
-		type: 'options',
-		options: [
-			{ name: 'Never', value: '', description: 'Do not autoplay (player default)' },
-			{ name: 'Always', value: 'always', description: 'Autoplay as soon as the player loads' },
-			{ name: 'On Scroll', value: 'on-scroll', description: 'Autoplay when the player scrolls into view' },
-		],
-		default: '',
-		description: 'When the video should start playing automatically. Most browsers require the player to be muted for autoplay to work.',
-		displayOptions: {
-			show: { resource: ['transform'], operation: ['videoPlayer'] },
-		},
-	},
-	{
-		displayName: 'Muted',
-		name: 'playerMuted',
-		type: 'boolean',
-		default: false,
-		description: 'Whether the player starts muted',
-		displayOptions: {
-			show: { resource: ['transform'], operation: ['videoPlayer'] },
-		},
-	},
-	{
-		displayName: 'Loop',
-		name: 'playerLoop',
-		type: 'boolean',
-		default: false,
-		description: 'Whether the video restarts from the beginning when it ends',
-		displayOptions: {
-			show: { resource: ['transform'], operation: ['videoPlayer'] },
-		},
-	},
-	{
-		displayName: 'Source Types',
-		name: 'playerSourceTypes',
-		type: 'multiOptions',
-		options: [
-			{ name: 'Auto', value: 'auto', description: 'Let the player pick the best available format' },
-			{ name: 'HLS', value: 'hls', description: 'HTTP Live Streaming (adaptive bitrate)' },
-			{ name: 'MP4', value: 'mp4', description: 'Progressive MP4' },
-			{ name: 'MPEG-DASH', value: 'dash', description: 'Dynamic Adaptive Streaming over HTTP' },
-			{ name: 'Ogg', value: 'ogg', description: 'Progressive Ogg' },
-			{ name: 'WebM', value: 'webm', description: 'Progressive WebM' },
-		],
-		default: [],
-		description: 'Preferred delivery formats in fallback order. Leave empty to use the player\'s default format selection.',
-		displayOptions: {
-			show: { resource: ['transform'], operation: ['videoPlayer'] },
-		},
-	},
-	{
-		displayName: 'Poster',
-		name: 'playerPoster',
-		type: 'string',
-		default: '',
-		description: 'URL of the still image shown before playback starts. Leave empty to use a frame from the video.',
-		displayOptions: {
-			show: { resource: ['transform'], operation: ['videoPlayer'] },
-		},
-	},
-	{
-		displayName: 'Transformation',
-		name: 'playerTransformation',
-		type: 'string',
-		default: '',
-		description: 'Cloudinary transformation string applied to the video source in the self-hosted player (player_config output). Note: not included in the iframe embed URL — the Cloudinary player iframe does not support video-level transformations via URL parameters.',
-		displayOptions: {
-			show: { resource: ['transform'], operation: ['videoPlayer'] },
-		},
-	},
-	{
-		displayName: 'Fluid',
-		name: 'playerFluid',
-		type: 'boolean',
-		default: false,
-		description: 'Whether the player resizes responsively to fill its container. When on, Width and Height are replaced by an aspect-ratio CSS style.',
-		displayOptions: {
-			show: { resource: ['transform'], operation: ['videoPlayer'] },
-		},
-	},
-	{
-		displayName: 'Width',
-		name: 'playerWidth',
-		type: 'number',
-		default: 0,
-		description: 'Player width in pixels. Leave 0 for the default (640 px).',
-		displayOptions: {
-			show: { resource: ['transform'], operation: ['videoPlayer'], playerFluid: [false] },
-		},
-	},
-	{
-		displayName: 'Height',
-		name: 'playerHeight',
-		type: 'number',
-		default: 0,
-		description: 'Player height in pixels. Leave 0 for the default (360 px).',
-		displayOptions: {
-			show: { resource: ['transform'], operation: ['videoPlayer'], playerFluid: [false] },
-		},
-	},
-	{
-		displayName: 'Aspect Ratio',
-		name: 'playerAspectRatio',
-		type: 'options',
-		options: [
-			{ name: '1:1', value: '1:1' },
-			{ name: '16:9', value: '16:9' },
-			{ name: '9:16', value: '9:16' },
-			{ name: 'Default', value: '' },
-		],
-		default: '',
-		description: 'Player aspect ratio. Leave unset to use the video\'s natural dimensions.',
-		displayOptions: {
-			show: { resource: ['transform'], operation: ['videoPlayer'] },
-		},
-	},
-	{
-		displayName: 'Skin',
-		name: 'playerSkin',
-		type: 'options',
-		options: [
-			{ name: 'Dark', value: 'dark' },
-			{ name: 'Light', value: 'light' },
-		],
-		default: 'dark',
-		description: 'Player theme',
-		displayOptions: {
-			show: { resource: ['transform'], operation: ['videoPlayer'] },
-		},
-	},
-	{
-		displayName: 'Base Color',
-		name: 'playerBaseColor',
-		type: 'color',
-		default: '',
-		description: 'Player base (background) color, as a hex value',
-		displayOptions: {
-			show: { resource: ['transform'], operation: ['videoPlayer'] },
-		},
-	},
-	{
-		displayName: 'Accent Color',
-		name: 'playerAccentColor',
-		type: 'color',
-		default: '',
-		description: 'Player accent (highlight) color, as a hex value',
-		displayOptions: {
-			show: { resource: ['transform'], operation: ['videoPlayer'] },
-		},
-	},
-	{
-		displayName: 'Text Color',
-		name: 'playerTextColor',
-		type: 'color',
-		default: '',
-		description: 'Player text color, as a hex value',
-		displayOptions: {
-			show: { resource: ['transform'], operation: ['videoPlayer'] },
-		},
-	},
-	{
-		displayName: 'Font Face',
-		name: 'playerFontFace',
-		type: 'string',
-		default: '',
-		description: 'The font applied to player text elements (titles, descriptions, recommendations, time counter). Accepts a Google Font name.',
-		displayOptions: {
-			show: { resource: ['transform'], operation: ['videoPlayer'] },
-		},
-	},
-	{
-		displayName: 'Advanced Player Options',
-		name: 'playerAdvancedOptions',
-		type: 'collection',
-		placeholder: 'Add Option',
-		default: {},
-		displayOptions: {
-			show: { resource: ['transform'], operation: ['videoPlayer'] },
-		},
-		options: [
-			{
-				displayName: 'Big Play Button',
-				name: 'bigPlayButton',
-				type: 'boolean',
-				default: true,
-				description: 'Whether to show a larger central play button when the video is paused',
-			},
-			{
-				displayName: 'Chapters Button',
-				name: 'chaptersButton',
-				type: 'boolean',
-				default: false,
-				description: 'Whether to show the chapters button, which opens a list of chapters',
-			},
-			{
-				displayName: 'Floating When Not Visible',
-				name: 'floatingWhenNotVisible',
-				type: 'options',
-				options: [
-					{ name: 'Disabled', value: 'disabled' },
-					{ name: 'Left', value: 'left' },
-					{ name: 'Right', value: 'right' },
-				],
-				default: 'disabled',
-				description: 'Whether the player should float into a mini-player corner when scrolled out of view',
-			},
-			{
-				displayName: 'HDR',
-				name: 'hdr',
-				type: 'boolean',
-				default: false,
-				description: 'Whether to deliver HDR video when the viewer\'s device and browser support it',
-			},
-			{
-				displayName: 'Picture In Picture Toggle',
-				name: 'pictureInPictureToggle',
-				type: 'boolean',
-				default: false,
-				description: 'Whether to show the picture-in-picture toggle, letting users view the video in a floating window',
-			},
-			{
-				displayName: 'Plays Inline',
-				name: 'playsinline',
-				type: 'boolean',
-				default: false,
-				description: 'Whether to prevent the player from entering fullscreen automatically on iOS when playback starts',
-			},
-			{
-				displayName: 'Seek Thumbnails',
-				name: 'seekThumbnails',
-				type: 'boolean',
-				default: true,
-				description: 'Whether to show thumbnail previews when seeking with the seek bar',
-			},
-			{
-				displayName: 'Show Controls',
-				name: 'controls',
-				type: 'boolean',
-				default: true,
-				description: 'Whether to show the built-in playback controls (play, pause, volume, fullscreen, etc.)',
-			},
-		],
-	},
-
-	// ── Multi-Step Transformation ────────────────────────────────────────────
+	// ── Combine Transformations (multi-step) ─────────────────────────────────
 	{
 		displayName: 'Resource Type',
 		name: 'multiStepResourceType',
@@ -661,7 +416,7 @@ export const transformFields: INodeProperties[] = [
 		default: 'image',
 		description: 'The type of asset the steps target. Determines the URL path and whether Optimize emits f_auto or f_auto:video.',
 		displayOptions: {
-			show: { resource: ['transform'], operation: ['multiStep'] },
+			show: { resource: ['transform'], operation: ['combineTransformations'] },
 		},
 	},
 	{
@@ -674,7 +429,7 @@ export const transformFields: INodeProperties[] = [
 		description:
 			'Transformation steps applied in order. Each step becomes one chained Cloudinary component (drag to reorder).',
 		displayOptions: {
-			show: { resource: ['transform'], operation: ['multiStep'] },
+			show: { resource: ['transform'], operation: ['combineTransformations'] },
 		},
 		options: [
 			{
