+++ package.json
@@ -46,9 +46,6 @@
       "dist/nodes/Cloudinary/Cloudinary.node.js"
     ]
   },
-  "dependencies": {
-    "@cloudinary/url-gen": "^1.22.0"
-  },
   "devDependencies": {
     "@types/node": "^24.1.0",
     "@typescript-eslint/parser": "~8.32.0",
